fix(pins): update pin.js implementation to match tests
